Moved module to seperate file
